the start of my replay-implementation all code related to it must be activated with compiler-switch HAVE_GHOST_REPLAY folder "replay" added to store the replay-files
there are 2 main classes. ReplayRecorder & ReplayPlayer. both are members of class World. the first is for recording/saving a replay and the latter one for showing the replay/loading/ghost-replay

*recording* a game works already very basic, only karts position-struct is stored and some data about the game (track,karts,player...)

the replay file currently is written to replay-folder in world's destructor to replay/test.rph
and it is loaded into ReplayPlayer in World's constructor

git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/trunk/supertuxkart@1243 178a84e3-b1eb-0310-8ba1-8eac791a3b58