xrandr: now it seems to work fine. I enable it for testing.
This should be useful for people which have two or more display devices. It changes resolution only on main screen and leaves other screens untouched.

Currently I use screen which is the most on left (or on top). Ideally we should be able to choose in settings which screen we want to use. But support it on all platforms is rather too much of work. We could also detect from which screen STK was executed. This can be done differently on different window managers, but nothing impossible. Maybe in future.

Perhaps we should add fallback using old xrandr method.

BTW. in some specific cases it works better than SDL2 ;-)