Make the kart classes differences in turn radius bigger
Now that the turn radius bug is fixed, these differences are more meaningful. The reference turn radius is close to the tighter numbers observed before, so this isn't actually a +15/30% to turn radius compared to previous effective numbers.

A turn radius of 19.5 at 25 of speed is still very manageable.

This doesn't preclude further tweaks in the future, once largely playtested in the beta.