Switch public ratings to a compound of raw score and RD
The raw score remain used in the core of the formula for expected scores computations.

Also improve the handling of disconnects on rating, with new RD floors
(and so public rating penalty) for recurring disconnects (>= 3 in the last 64 races).
The raw score is less penalized than before, as it's not needed and
make expected score computations more reliable.