UI/UX: Put OK/Confirm/Yes on the right
In UX, back/cancel/etc buttons usually go on the left, while forward/confirm/etc go on the right.  (Note that even the icon for "Back" is a leftwards-pointing arrow.)  STK put the Back arrow button on general screens in the upper left corner, but before this commit, put back/cancel/etc on the right in dialogs and such.  This commit fixes things to be in line with standard UX conventions, and also in line with the pre-existing upper-left corner Back button placement on screens in STK.

https://ux.stackexchange.com/a/63239