Get rid of repeated code, add comments
easySQLQuery was generalized to return output rows if requested,
which allowed to shorten the code for many queries.

The code for binding values to sqlite statement was also repeated
many times. Two auxiliary structures were introduced, so that it's
possible to provide at the same time both those parameters which
require and those which don't require binding, in a single
StringUtils::insertValues() call.