Fix #5014
The TextBoxWidget relies under the hood on IGUIEditBox and behaves in a peculiar way. How exactly both the 'send' button and the chat box could be simultaneously active is unclear beyond involving use of both keyboard and mouse, and the underlying issue will likely remain until a deeper overhaul.

In the meantime, simply selecting back the chatbox after the send box has been triggered ensures that at most one line will be affected, and is convenient in other situations where the send box is used.