Fix a segfault when closing the game
Powerup objects needed to clean up audio sources because they were created and deleted for each race. The powerup audio object is created once, and reused until the program is closed, so there is no risk of memory leak. Trying to delete SFX sources in the object destructor is unnecessary.