Put karts upright quicker when only some wheels touch the ground (fix #5460)
Just applying a naïve corrective torque when some wheels are off the ground would lead to many unwelcome side effects, so a more accurate approach is needed.

The code's comments explain the procedure in more details, but in general this commit:
- Moves the logic applying the correcting torque to a dedicated function (this allows an early return if conditions are not met, and there is a lot more code there than before)
- Adds logic to evaluate if correcting the kart would help or not. To do this, we compare (1) the flying wheels height off the ground, (2) their simulated height off the ground with an upright kart and the wheels touching ground staying in the same position. A correcting torque is only applied if it would bring the flying wheels noticeably closer to the ground.
- Creates a new uprightRayCast function that raycasts in the direction gravity pulls the kart towards. It is important to make consistent and comparable raycasts to measure height with. Unlike the others in btKart, this raycast function also avoids changing state.

As a result, the duration the kart spends with only two wheels on the ground is reduced by over 70% in the test track Sven provided. There is therefore still a small period with the kart uncontrollable (making the kart rotate too fast would look strange), but the player can quickly recover control and avoid a crash or having to rescue.

Because the torque is restricted to the roll axis, this commit only addresses the case of wheels on one side (left or right) being on the ground while wheels on the other sides are not. The detection logic works correctly when the pairings are front/back instead, there is simply no torque on the pitch axis. Because the conditions for this situation are very specific and don't arise in normal gameplay, I have left this as-is.