Increase turn radius in lower difficulties or with handicap
If a difficulty or handicap is X times slower than the reference speed, the speed factor for turn radius is approximately sqrt(X). As a result, the different turning capabilities of various kart classes remain more relevant with those settings, but it remains true that: the lower the difficulty, the easier it is to turn when at maximum speed.