Play the parachute sound if a local player is affected
The transfer to PowerupAudio fixed many issues, but it created a regression for the parachute sound: previously the parachute sound was also played if a local player was a target of the parachute powerup, after the PowerupAudio change no sound would be played in that situation. This commit restores previous behavior.

This also fixes a compilation warning related to this code.