Only apply the increased turn radius in lower difficulties above a threshold speed
We use a lowerbound value at which the increased turn radius starts being applied and an upperbound value above which it is fully applied. This change allows to preserve the intended racing behavior (turn radius characteristics of karts keep relevance at lower difficulties or with handicap, as without the speed factor heavies are obviously superior) while significantly improving low-speed maneuvers (used especially in battle and soccer) that had become too sluggish as a side-effect.

Also:
- Add error logs to catch misuse of the lowerbound and upperbound values
- Fix a comparison that was mistakenly written as an assignation in the previous commit.