Add support for the false start animation
If a player gets a start penalty, the false start animation will play once the countdown finishes.

A placeholder enum value is used so that assumptions about the end frame being 2 places further in the enum hold true. While not the most elegant solution, it's very simple, and looping is automatically prevented by the placeholder frame animation not being set.