Remove all the code for terrain impulses
This code was supposed to help improve the outcome of karts colliding against track obstacles, but it didn't work properly.

The 'normal' impulse type was already de facto disabled, as the game was applying an impulse of duration 0 and 0 strength. Meanwhile, the 'impulse to driveline' was a fundamentally flawed concept as it came with unescapable edge cases, and it was not enabled either (since stk_config was set to use the normal impulse).