Move more functions to GraphicalPresets
The function to probe the currently active graphical preset, to set a given graphical preset, to probe the current ImageQuality level, and to set an ImageQuality level in config have also been transferred.

Like the previous commit, this will be helpful in contexts other than the video settings screen, for example applying a graphical preset using a command line option.