Further tweaks for nitro
While long duration naive usage got a slightly lower penalty compared to proper burst usage, a more detailed analysis indicates that the details are particularly dependent on the used kart class and the discrete values received.

The closest the minimum burst consumption was to a single small nitro, and the smaller the gap between naive usage and burst usage when nitro was only collected in small cans and used quickly thereafter.

As a consequence, with better efficiency, the previous commit actually increased the gap in the single-can case. For heavy karts this was warranted, as the previous difference of about ~14% was too small, but for light karts the increase from ~43% to ~52% was unwanted.

Therefore:
- The min-burst duration is slightly increased for medium and light karts (10% and 20% respectively). This brings the minimum consumption per burst closer between kart classes (from 0.5560 for lights to 0.6435 for heavies).
- The base consumption per second is reduced from 0.72 to 0.7
- Base consumption is normalized to heavy karts instead of light karts (0.7 * 1.67 ~= 1.17).
- Because longer minimum bursts mean less efficiency for burst usage, the max speed boost from nitro is increased for medium and light karts (from 10% / 20% to 15% / 30%) without a commensurate reduction in base efficiency.

As a result of these changes, the nitro efficiency gap between heavies and other kart classes is slightly reduced in the ideal case. However, in the naive (no-burst) usage case with only a small nitro, the gap is bigger than before.

The following tables give the approximate effiency advantage of burst usage (the precise value of the fade-out time giving speed but no engine force boosts is difficult to determine, and the time needed to accelerate to an increased max speed is not taken into account) for various kart classes and nitro quantities:

(1 nitro) | 1.5  | Evo (old) | Evo (new)
Light     | ~58% |      ~43% |     ~37%
Medium    | ~35% |      ~30% |     ~36%
Heavy     |  ~9% |      ~14% |     ~30%

(2.5 nitros) |  1.5  | Evo (old) | Evo (new)
Light        | ~162% |     ~132% |     ~90%
Medium       | ~136% |     ~118% |     ~93%
Heavy        | ~127% |      ~99% |     ~94%

(5 nitros) |  1.5  | Evo (old) | Evo (new)
Light      | ~236% |     ~192% |    ~118%
Medium     | ~214% |     ~181% |    ~125%
Heavy      | ~206% |     ~165% |    ~131%